dateFn utils분리, 이미지 크기 수정
