+++ hanarowa/src/utils/facility.ts
@@ -1,0 +1,15 @@
+const createFacilityDate = () => {
+  const today = new Date();
+  const tomorrow = new Date(today);
+  tomorrow.setDate(today.getDate() + 1);
+
+  const formatDate = (date: Date): string => {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0');
+    const day = String(date.getDate()).padStart(2, '0');
+    return `${year}-${month}-${day}`;
+  };
+  return { formatDate, tomorrow, today };
+};
+
+export default createFacilityDate;
